sub GameService
