+++ src/main/java/com/likelion/voca/VocaApplication.java
@@ -12,6 +12,7 @@ public class VocaApplication {
 
 
 	public static void main(String[] args) {
+		/*
 		try {
 			Class.forName("com.mysql.cj.jdbc.Driver");
 			Connection conn = DriverManager.getConnection("jdbc:mysql://127.0.0.1:3306/vocaboard");
@@ -23,7 +24,7 @@ public class VocaApplication {
 		} catch(SQLException error) {
 			System.out.println("Connection error");
 		}
-
+		*/
 		SpringApplication.run(VocaApplication.class, args);
 	}
 }
